[refactor]: extract some logic from AuthStoreBackend
