+++ xline/src/storage/auth_store/store.rs
@@ -1,19 +1,51 @@
-use std::{collections::VecDeque, sync::Arc};
+use std::{
+    cmp::Ordering,
+    collections::{HashMap, VecDeque},
+    sync::{
+        atomic::{AtomicBool, Ordering as AtomicOrdering},
+        Arc,
+    },
+};
 
 use anyhow::Result;
+use clippy_utilities::Cast;
+use itertools::Itertools;
 use jsonwebtoken::{DecodingKey, EncodingKey};
+use log::debug;
+use parking_lot::RwLock;
+use pbkdf2::{
+    password_hash::{PasswordHash, PasswordVerifier},
+    Pbkdf2,
+};
 use tokio::sync::mpsc;
+use utils::parking_lot_lock::RwLockMap;
 
-use super::backend::{ROOT_ROLE, ROOT_USER};
+use super::{
+    backend::{ROOT_ROLE, ROOT_USER},
+    perms::{JwtTokenManager, PermissionCache, TokenClaims, TokenOperate, UserPermissions},
+};
 use crate::{
     header_gen::HeaderGenerator,
     rpc::{
-        DeleteRangeRequest, LeaseRevokeRequest, PutRequest, RangeRequest, Request, RequestOp,
-        RequestWithToken, RequestWrapper, TxnRequest, Type,
+        AuthDisableRequest, AuthDisableResponse, AuthEnableRequest, AuthEnableResponse,
+        AuthRoleAddRequest, AuthRoleAddResponse, AuthRoleDeleteRequest, AuthRoleDeleteResponse,
+        AuthRoleGetRequest, AuthRoleGetResponse, AuthRoleGrantPermissionRequest,
+        AuthRoleGrantPermissionResponse, AuthRoleListRequest, AuthRoleListResponse,
+        AuthRoleRevokePermissionRequest, AuthRoleRevokePermissionResponse, AuthStatusRequest,
+        AuthStatusResponse, AuthUserAddRequest, AuthUserAddResponse, AuthUserChangePasswordRequest,
+        AuthUserChangePasswordResponse, AuthUserDeleteRequest, AuthUserDeleteResponse,
+        AuthUserGetRequest, AuthUserGetResponse, AuthUserGrantRoleRequest,
+        AuthUserGrantRoleResponse, AuthUserListRequest, AuthUserListResponse,
+        AuthUserRevokeRoleRequest, AuthUserRevokeRoleResponse, AuthenticateRequest,
+        AuthenticateResponse, DeleteRangeRequest, LeaseRevokeRequest, Permission, PutRequest,
+        RangeRequest, Request, RequestOp, RequestWithToken, RequestWrapper, Role, TxnRequest, Type,
+        User,
     },
     server::command::{CommandResponse, KeyRange, SyncResponse},
     storage::{
-        auth_store::backend::AuthStoreBackend, lease_store::LeaseMessage, storage_api::StorageApi,
+        auth_store::backend::AuthStoreBackend,
+        lease_store::{Lease, LeaseMessage},
+        storage_api::StorageApi,
         ExecuteError,
     },
 };
@@ -26,6 +58,17 @@ where
 {
     /// Auth store Backend
     inner: Arc<AuthStoreBackend<S>>,
+    /// Enabled
+    enabled: AtomicBool,
+    /// Lease command sender
+    lease_cmd_tx: mpsc::Sender<LeaseMessage>,
+    /// Header generator
+    header_gen: Arc<HeaderGenerator>,
+    /// Permission cache
+    permission_cache: RwLock<PermissionCache>,
+
+    /// The manager of token
+    token_manager: Option<JwtTokenManager>,
 }
 
 impl<S> AuthStore<S>
@@ -41,13 +84,95 @@ where
         storage: Arc<S>,
     ) -> Self {
         Self {
-            inner: Arc::new(AuthStoreBackend::new(
-                lease_cmd_tx,
-                key_pair,
-                header_gen,
-                storage,
-            )),
+            inner: Arc::new(AuthStoreBackend::new(storage)),
+            enabled: AtomicBool::new(false),
+            lease_cmd_tx,
+            header_gen,
+            permission_cache: RwLock::new(PermissionCache::new()),
+            token_manager: key_pair.map(|(encoding_key, decoding_key)| {
+                JwtTokenManager::new(encoding_key, decoding_key)
+            }),
         }
+    }
+
+    /// Get Lease by lease id
+    async fn get_lease(&self, lease_id: i64) -> Option<Lease> {
+        let (detach, rx) = LeaseMessage::look_up(lease_id);
+        assert!(
+            self.lease_cmd_tx.send(detach).await.is_ok(),
+            "lease_cmd_tx is closed"
+        );
+        rx.await.unwrap_or_else(|_e| panic!("res sender is closed"))
+    }
+
+    /// Get enabled of Auth store
+    pub(super) fn is_enabled(&self) -> bool {
+        self.enabled.load(AtomicOrdering::Relaxed)
+    }
+
+    /// Assign token
+    pub(crate) fn assign(&self, username: &str) -> Result<String, ExecuteError> {
+        match self.token_manager {
+            Some(ref token_manager) => token_manager
+                .assign(username, self.revision())
+                .map_err(|_ignore| ExecuteError::invalid_auth_token()),
+            None => Err(ExecuteError::token_manager_not_init()),
+        }
+    }
+
+    /// verify token
+    pub(crate) fn verify_token(&self, token: &str) -> Result<TokenClaims, ExecuteError> {
+        match self.token_manager {
+            Some(ref token_manager) => token_manager
+                .verify(token)
+                .map_err(|_ignore| ExecuteError::invalid_auth_token()),
+            None => Err(ExecuteError::token_manager_not_init()),
+        }
+    }
+
+    /// create permission cache
+    fn create_permission_cache(&self) -> Result<(), ExecuteError> {
+        let mut permission_cache = PermissionCache::new();
+        for user in self.inner.get_all_users()? {
+            let user_permission = self.get_user_permissions(&user);
+            let username = String::from_utf8_lossy(&user.name).to_string();
+            let _ignore = permission_cache
+                .user_permissions
+                .insert(username, user_permission);
+        }
+        self.permission_cache
+            .map_write(|mut cache| *cache = permission_cache);
+        Ok(())
+    }
+
+    /// get user permissions
+    pub(crate) fn get_user_permissions(&self, user: &User) -> UserPermissions {
+        let mut user_permission = UserPermissions::new();
+        for role_name in &user.roles {
+            let Ok(role) = self.inner.get_role(role_name) else {
+                continue;
+            };
+            for permission in role.key_permission {
+                let key_range = KeyRange {
+                    start: permission.key,
+                    end: permission.range_end,
+                };
+                #[allow(clippy::unwrap_used)] // safe unwrap
+                match Type::from_i32(permission.perm_type).unwrap() {
+                    Type::Readwrite => {
+                        user_permission.read.push(key_range.clone());
+                        user_permission.write.push(key_range.clone());
+                    }
+                    Type::Write => {
+                        user_permission.write.push(key_range.clone());
+                    }
+                    Type::Read => {
+                        user_permission.read.push(key_range.clone());
+                    }
+                }
+            }
+        }
+        user_permission
     }
 
     /// execute a auth request
@@ -55,9 +180,335 @@ where
         &self,
         request: &RequestWithToken,
     ) -> Result<CommandResponse, ExecuteError> {
-        self.inner
-            .handle_auth_req(&request.request)
-            .map(CommandResponse::new)
+        #[allow(clippy::wildcard_enum_match_arm)]
+        let res = match request.request {
+            RequestWrapper::AuthEnableRequest(ref req) => {
+                self.handle_auth_enable_request(req).map(Into::into)
+            }
+            RequestWrapper::AuthDisableRequest(ref req) => {
+                Ok(self.handle_auth_disable_request(req).into())
+            }
+            RequestWrapper::AuthStatusRequest(ref req) => {
+                Ok(self.handle_auth_status_request(req).into())
+            }
+            RequestWrapper::AuthUserAddRequest(ref req) => {
+                self.handle_user_add_request(req).map(Into::into)
+            }
+            RequestWrapper::AuthUserGetRequest(ref req) => {
+                self.handle_user_get_request(req).map(Into::into)
+            }
+            RequestWrapper::AuthUserListRequest(ref req) => {
+                self.handle_user_list_request(req).map(Into::into)
+            }
+            RequestWrapper::AuthUserGrantRoleRequest(ref req) => {
+                self.handle_user_grant_role_request(req).map(Into::into)
+            }
+            RequestWrapper::AuthUserRevokeRoleRequest(ref req) => {
+                self.handle_user_revoke_role_request(req).map(Into::into)
+            }
+            RequestWrapper::AuthUserChangePasswordRequest(ref req) => self
+                .handle_user_change_password_request(req)
+                .map(Into::into),
+            RequestWrapper::AuthUserDeleteRequest(ref req) => {
+                self.handle_user_delete_request(req).map(Into::into)
+            }
+            RequestWrapper::AuthRoleAddRequest(ref req) => {
+                self.handle_role_add_request(req).map(Into::into)
+            }
+            RequestWrapper::AuthRoleGetRequest(ref req) => {
+                self.handle_role_get_request(req).map(Into::into)
+            }
+            RequestWrapper::AuthRoleGrantPermissionRequest(ref req) => self
+                .handle_role_grant_permission_request(req)
+                .map(Into::into),
+            RequestWrapper::AuthRoleRevokePermissionRequest(ref req) => self
+                .handle_role_revoke_permission_request(req)
+                .map(Into::into),
+            RequestWrapper::AuthRoleDeleteRequest(ref req) => {
+                self.handle_role_delete_request(req).map(Into::into)
+            }
+            RequestWrapper::AuthRoleListRequest(ref req) => {
+                self.handle_role_list_request(req).map(Into::into)
+            }
+            RequestWrapper::AuthenticateRequest(ref req) => {
+                self.handle_authenticate_request(req).map(Into::into)
+            }
+            _ => {
+                unreachable!("Other request should not be sent to this store");
+            }
+        };
+        res.map(CommandResponse::new)
+    }
+
+    /// Handle `AuthEnableRequest`
+    fn handle_auth_enable_request(
+        &self,
+        _req: &AuthEnableRequest,
+    ) -> Result<AuthEnableResponse, ExecuteError> {
+        debug!("handle_auth_enable");
+        let res = Ok(AuthEnableResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+        });
+        if self.is_enabled() {
+            debug!("auth is already enabled");
+            return res;
+        }
+        let user = self.inner.get_user(ROOT_USER)?;
+        if user.roles.binary_search(&ROOT_ROLE.to_owned()).is_err() {
+            return Err(ExecuteError::root_role_not_exist());
+        }
+        res
+    }
+
+    /// Handle `AuthDisableRequest`
+    fn handle_auth_disable_request(&self, _req: &AuthDisableRequest) -> AuthDisableResponse {
+        debug!("handle_auth_disable");
+        if !self.is_enabled() {
+            debug!("auth is already disabled");
+        }
+        AuthDisableResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+        }
+    }
+
+    /// Handle `AuthStatusRequest`
+    fn handle_auth_status_request(&self, _req: &AuthStatusRequest) -> AuthStatusResponse {
+        debug!("handle_auth_status");
+        AuthStatusResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+            auth_revision: self.revision().cast(),
+            enabled: self.is_enabled(),
+        }
+    }
+
+    /// Handle `AuthenticateRequest`
+    fn handle_authenticate_request(
+        &self,
+        req: &AuthenticateRequest,
+    ) -> Result<AuthenticateResponse, ExecuteError> {
+        debug!("handle_authenticate_request");
+        if !self.is_enabled() {
+            return Err(ExecuteError::auth_not_enabled());
+        }
+        let token = self.assign(&req.name)?;
+        Ok(AuthenticateResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+            token,
+        })
+    }
+
+    /// Handle `AuthUserAddRequest`
+    fn handle_user_add_request(
+        &self,
+        req: &AuthUserAddRequest,
+    ) -> Result<AuthUserAddResponse, ExecuteError> {
+        debug!("handle_user_add_request");
+        if self.inner.get_user(&req.name).is_ok() {
+            return Err(ExecuteError::user_already_exists(&req.name));
+        }
+        Ok(AuthUserAddResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+        })
+    }
+
+    /// Handle `AuthUserGetRequest`
+    fn handle_user_get_request(
+        &self,
+        req: &AuthUserGetRequest,
+    ) -> Result<AuthUserGetResponse, ExecuteError> {
+        debug!("handle_user_add_request");
+        let user = self.inner.get_user(&req.name)?;
+        Ok(AuthUserGetResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+            roles: user.roles,
+        })
+    }
+
+    /// Handle `AuthUserListRequest`
+    fn handle_user_list_request(
+        &self,
+        _req: &AuthUserListRequest,
+    ) -> Result<AuthUserListResponse, ExecuteError> {
+        debug!("handle_user_list_request");
+        let users = self
+            .inner
+            .get_all_users()?
+            .into_iter()
+            .map(|u| String::from_utf8_lossy(&u.name).to_string())
+            .collect();
+        Ok(AuthUserListResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+            users,
+        })
+    }
+
+    /// Handle `AuthUserDeleteRequest`
+    fn handle_user_delete_request(
+        &self,
+        req: &AuthUserDeleteRequest,
+    ) -> Result<AuthUserDeleteResponse, ExecuteError> {
+        debug!("handle_user_delete_request");
+        if self.is_enabled() && (req.name == ROOT_USER) {
+            return Err(ExecuteError::invalid_auth_management());
+        }
+        let _user = self.inner.get_user(&req.name)?;
+        Ok(AuthUserDeleteResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+        })
+    }
+
+    /// Handle `AuthUserChangePasswordRequest`
+    fn handle_user_change_password_request(
+        &self,
+        req: &AuthUserChangePasswordRequest,
+    ) -> Result<AuthUserChangePasswordResponse, ExecuteError> {
+        debug!("handle_user_change_password_request");
+        let user = self.inner.get_user(&req.name)?;
+        let need_password = user.options.as_ref().map_or(true, |o| !o.no_password);
+        if need_password && req.hashed_password.is_empty() {
+            return Err(ExecuteError::no_password_user());
+        }
+        Ok(AuthUserChangePasswordResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+        })
+    }
+
+    /// Handle `AuthUserGrantRoleRequest`
+    fn handle_user_grant_role_request(
+        &self,
+        req: &AuthUserGrantRoleRequest,
+    ) -> Result<AuthUserGrantRoleResponse, ExecuteError> {
+        debug!("handle_user_grant_role_request");
+        let _user = self.inner.get_user(&req.user)?;
+        if req.role != ROOT_ROLE {
+            let _role = self.inner.get_role(&req.role)?;
+        }
+        Ok(AuthUserGrantRoleResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+        })
+    }
+
+    /// Handle `AuthUserRevokeRoleRequest`
+    fn handle_user_revoke_role_request(
+        &self,
+        req: &AuthUserRevokeRoleRequest,
+    ) -> Result<AuthUserRevokeRoleResponse, ExecuteError> {
+        debug!("handle_user_revoke_role_request");
+        if self.is_enabled() && (req.name == ROOT_USER) && (req.role == ROOT_ROLE) {
+            return Err(ExecuteError::invalid_auth_management());
+        }
+        let user = self.inner.get_user(&req.name)?;
+        if user.roles.binary_search(&req.role).is_err() {
+            return Err(ExecuteError::role_not_granted(&req.role));
+        }
+        Ok(AuthUserRevokeRoleResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+        })
+    }
+
+    /// Handle `AuthRoleAddRequest`
+    fn handle_role_add_request(
+        &self,
+        req: &AuthRoleAddRequest,
+    ) -> Result<AuthRoleAddResponse, ExecuteError> {
+        debug!("handle_role_add_request");
+        if self.inner.get_role(&req.name).is_ok() {
+            return Err(ExecuteError::role_already_exists(&req.name));
+        }
+        Ok(AuthRoleAddResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+        })
+    }
+
+    /// Handle `AuthRoleGetRequest`
+    fn handle_role_get_request(
+        &self,
+        req: &AuthRoleGetRequest,
+    ) -> Result<AuthRoleGetResponse, ExecuteError> {
+        debug!("handle_role_get_request");
+        let role = self.inner.get_role(&req.role)?;
+        let perm = if role.name == ROOT_ROLE.as_bytes() {
+            vec![Permission {
+                #[allow(clippy::as_conversions)] // This cast is always valid
+                perm_type: Type::Readwrite as i32,
+                key: vec![],
+                range_end: vec![0],
+            }]
+        } else {
+            role.key_permission
+        };
+        Ok(AuthRoleGetResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+            perm,
+        })
+    }
+
+    /// Handle `AuthRoleListRequest`
+    fn handle_role_list_request(
+        &self,
+        _req: &AuthRoleListRequest,
+    ) -> Result<AuthRoleListResponse, ExecuteError> {
+        debug!("handle_role_list_request");
+        let roles = self
+            .inner
+            .get_all_roles()?
+            .into_iter()
+            .map(|r| String::from_utf8_lossy(&r.name).to_string())
+            .collect();
+        Ok(AuthRoleListResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+            roles,
+        })
+    }
+
+    /// Handle `UserRoleDeleteRequest`
+    fn handle_role_delete_request(
+        &self,
+        req: &AuthRoleDeleteRequest,
+    ) -> Result<AuthRoleDeleteResponse, ExecuteError> {
+        debug!("handle_role_delete_request");
+        if self.is_enabled() && req.role == ROOT_ROLE {
+            return Err(ExecuteError::invalid_auth_management());
+        }
+        let _role = self.inner.get_role(&req.role)?;
+        Ok(AuthRoleDeleteResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+        })
+    }
+
+    /// Handle `AuthRoleGrantPermissionRequest`
+    fn handle_role_grant_permission_request(
+        &self,
+        req: &AuthRoleGrantPermissionRequest,
+    ) -> Result<AuthRoleGrantPermissionResponse, ExecuteError> {
+        debug!("handle_role_grant_permission_request");
+        let _role = self.inner.get_role(&req.name)?;
+        Ok(AuthRoleGrantPermissionResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+        })
+    }
+
+    /// Handle `AuthRoleRevokePermissionRequest`
+    fn handle_role_revoke_permission_request(
+        &self,
+        req: &AuthRoleRevokePermissionRequest,
+    ) -> Result<AuthRoleRevokePermissionResponse, ExecuteError> {
+        debug!("handle_role_revoke_permission_request");
+        let role = self.inner.get_role(&req.role)?;
+        if role
+            .key_permission
+            .binary_search_by(|p| match p.key.cmp(&req.key) {
+                Ordering::Equal => p.range_end.cmp(&req.range_end),
+                Ordering::Less => Ordering::Less,
+                Ordering::Greater => Ordering::Greater,
+            })
+            .is_err()
+        {
+            return Err(ExecuteError::permission_not_granted());
+        }
+        Ok(AuthRoleRevokePermissionResponse {
+            header: Some(self.header_gen.gen_header_without_revision()),
+        })
     }
 
     /// sync a auth request
@@ -65,9 +516,328 @@ where
         &self,
         request: &RequestWithToken,
     ) -> Result<SyncResponse, ExecuteError> {
-        self.inner
-            .sync_request(&request.request)
-            .map(SyncResponse::new)
+        #[allow(clippy::wildcard_enum_match_arm)]
+        match request.request {
+            RequestWrapper::AuthEnableRequest(ref req) => {
+                debug!("Sync AuthEnableRequest {:?}", req);
+                self.sync_auth_enable_request(req)?;
+            }
+            RequestWrapper::AuthDisableRequest(ref req) => {
+                debug!("Sync AuthDisableRequest {:?}", req);
+                self.sync_auth_disable_request(req)?;
+            }
+            RequestWrapper::AuthStatusRequest(ref req) => {
+                debug!("Sync AuthStatusRequest {:?}", req);
+            }
+            RequestWrapper::AuthUserAddRequest(ref req) => {
+                debug!("Sync AuthUserAddRequest {:?}", req);
+                self.sync_user_add_request(req)?;
+            }
+            RequestWrapper::AuthUserGetRequest(ref req) => {
+                debug!("Sync AuthUserGetRequest {:?}", req);
+            }
+            RequestWrapper::AuthUserListRequest(ref req) => {
+                debug!("Sync AuthUserListRequest {:?}", req);
+            }
+            RequestWrapper::AuthUserGrantRoleRequest(ref req) => {
+                debug!("Sync AuthUserGrantRoleRequest {:?}", req);
+                self.sync_user_grant_role_request(req)?;
+            }
+            RequestWrapper::AuthUserRevokeRoleRequest(ref req) => {
+                debug!("Sync AuthUserRevokeRoleRequest {:?}", req);
+                self.sync_user_revoke_role_request(req)?;
+            }
+            RequestWrapper::AuthUserChangePasswordRequest(ref req) => {
+                debug!("Sync AuthUserChangePasswordRequest {:?}", req);
+                self.sync_user_change_password_request(req)?;
+            }
+            RequestWrapper::AuthUserDeleteRequest(ref req) => {
+                debug!("Sync AuthUserDeleteRequest {:?}", req);
+                self.sync_user_delete_request(req)?;
+            }
+            RequestWrapper::AuthRoleAddRequest(ref req) => {
+                debug!("Sync AuthRoleAddRequest {:?}", req);
+                self.sync_role_add_request(req)?;
+            }
+            RequestWrapper::AuthRoleGetRequest(ref req) => {
+                debug!("Sync AuthRoleGetRequest {:?}", req);
+            }
+            RequestWrapper::AuthRoleGrantPermissionRequest(ref req) => {
+                debug!("Sync AuthRoleGrantPermissionRequest {:?}", req);
+                self.sync_role_grant_permission_request(req)?;
+            }
+            RequestWrapper::AuthRoleRevokePermissionRequest(ref req) => {
+                debug!("Sync AuthRoleRevokePermissionRequest {:?}", req);
+                self.sync_role_revoke_permission_request(req)?;
+            }
+            RequestWrapper::AuthRoleListRequest(ref req) => {
+                debug!("Sync AuthRoleListRequest {:?}", req);
+            }
+            RequestWrapper::AuthRoleDeleteRequest(ref req) => {
+                debug!("Sync AuthRoleDeleteRequest {:?}", req);
+                self.sync_role_delete_request(req)?;
+            }
+            RequestWrapper::AuthenticateRequest(ref req) => {
+                debug!("Sync AuthenticateRequest {:?}", req);
+            }
+            _ => {
+                unreachable!("Other request should not be sent to this store");
+            }
+        }
+        Ok(SyncResponse::new(self.header_gen.revision()))
+    }
+
+    /// Sync `AuthEnableRequest` and return whether authstore is changed.
+    fn sync_auth_enable_request(&self, _req: &AuthEnableRequest) -> Result<(), ExecuteError> {
+        if self.is_enabled() {
+            return Ok(());
+        }
+        self.inner.put_auth_enable(true)?;
+        self.enabled.store(true, AtomicOrdering::Relaxed);
+        self.create_permission_cache()
+    }
+
+    /// Sync `AuthDisableRequest` and return whether authstore is changed.
+    fn sync_auth_disable_request(&self, _req: &AuthDisableRequest) -> Result<(), ExecuteError> {
+        if !self.is_enabled() {
+            return Ok(());
+        }
+        self.inner.put_auth_enable(false)?;
+        self.enabled.store(false, AtomicOrdering::Relaxed);
+        Ok(())
+    }
+
+    /// Sync `AuthUserAddRequest` and return whether authstore is changed.
+    fn sync_user_add_request(&self, req: &AuthUserAddRequest) -> Result<(), ExecuteError> {
+        let user = User {
+            name: req.name.as_str().into(),
+            password: req.hashed_password.as_str().into(),
+            options: req.options.clone(),
+            roles: Vec::new(),
+        };
+        self.inner.put_user(&user)
+    }
+
+    /// Sync `AuthUserDeleteRequest` and return whether authstore is changed.
+    fn sync_user_delete_request(&self, req: &AuthUserDeleteRequest) -> Result<(), ExecuteError> {
+        self.inner.delete_user(&req.name)?;
+        self.permission_cache.map_write(|mut cache| {
+            let _ignore = cache.user_permissions.remove(&req.name);
+            cache.role_to_users_map.iter_mut().for_each(|(_, users)| {
+                if let Some((idx, _)) = users.iter().find_position(|uname| uname == &&req.name) {
+                    let _old = users.swap_remove(idx);
+                };
+            });
+        });
+        Ok(())
+    }
+
+    /// Sync `AuthUserChangePasswordRequest` and return whether authstore is changed.
+    fn sync_user_change_password_request(
+        &self,
+        req: &AuthUserChangePasswordRequest,
+    ) -> Result<(), ExecuteError> {
+        let mut user = self.inner.get_user(&req.name)?;
+        user.password = req.hashed_password.as_str().into();
+        self.inner.put_user(&user)
+    }
+
+    /// Sync `AuthUserGrantRoleRequest` and return whether authstore is changed.
+    fn sync_user_grant_role_request(
+        &self,
+        req: &AuthUserGrantRoleRequest,
+    ) -> Result<(), ExecuteError> {
+        let mut user = self.inner.get_user(&req.user)?;
+        let role = self.inner.get_role(&req.role);
+        if (req.role != ROOT_ROLE) && role.is_err() {
+            return Err(ExecuteError::role_not_found(&req.role));
+        }
+        let Err(idx) =  user.roles.binary_search(&req.role) else {
+            return Err(ExecuteError::user_already_has_role(&req.user, &req.role));
+        };
+        user.roles.insert(idx, req.role.clone());
+        self.inner.put_user(&user)?;
+        if let Ok(role) = role {
+            let perms = role.key_permission;
+            self.permission_cache.map_write(|mut cache| {
+                let entry = cache
+                    .user_permissions
+                    .entry(req.user.clone())
+                    .or_insert_with(UserPermissions::new);
+                for perm in perms {
+                    let key_range = KeyRange::new(perm.key, perm.range_end);
+                    #[allow(clippy::unwrap_used)] // safe unwrap
+                    match Type::from_i32(perm.perm_type).unwrap() {
+                        Type::Readwrite => {
+                            entry.read.push(key_range.clone());
+                            entry.write.push(key_range);
+                        }
+                        Type::Write => {
+                            entry.write.push(key_range);
+                        }
+                        Type::Read => {
+                            entry.read.push(key_range);
+                        }
+                    }
+                }
+                cache
+                    .role_to_users_map
+                    .entry(req.role.clone())
+                    .or_insert_with(Vec::new)
+                    .push(req.user.clone());
+            });
+        }
+        Ok(())
+    }
+
+    /// Sync `AuthUserRevokeRoleRequest` and return whether authstore is changed.
+    fn sync_user_revoke_role_request(
+        &self,
+        req: &AuthUserRevokeRoleRequest,
+    ) -> Result<(), ExecuteError> {
+        let mut user = self.inner.get_user(&req.name)?;
+        let idx = user
+            .roles
+            .binary_search(&req.role)
+            .map_err(|_ignore| ExecuteError::role_not_granted(&req.role))?;
+        let _ignore = user.roles.remove(idx);
+        self.permission_cache.map_write(|mut cache| {
+            let user_permissions = self.get_user_permissions(&user);
+            let _entry = cache
+                .role_to_users_map
+                .entry(req.role.clone())
+                .and_modify(|users| {
+                    if let Some((i, _)) = users.iter().find_position(|uname| uname == &&req.name) {
+                        let _old = users.swap_remove(i);
+                    };
+                });
+            let _old = cache
+                .user_permissions
+                .insert(req.name.clone(), user_permissions);
+        });
+        Ok(())
+    }
+
+    /// Sync `AuthRoleAddRequest` and return whether authstore is changed.
+    fn sync_role_add_request(&self, req: &AuthRoleAddRequest) -> Result<(), ExecuteError> {
+        let role = Role {
+            name: req.name.as_str().into(),
+            key_permission: Vec::new(),
+        };
+        self.inner.put_role(&role)
+    }
+
+    /// Sync `AuthRoleDeleteRequest` and return whether authstore is changed.
+    fn sync_role_delete_request(&self, req: &AuthRoleDeleteRequest) -> Result<(), ExecuteError> {
+        self.inner.delete_role(&req.role)?;
+        let users = self.inner.get_all_users()?;
+        let mut new_perms = HashMap::new();
+        for mut user in users {
+            if let Ok(idx) = user.roles.binary_search(&req.role) {
+                let _ignore = user.roles.remove(idx);
+                let perms = self.get_user_permissions(&user);
+                let _old = new_perms.insert(String::from_utf8_lossy(&user.name).to_string(), perms);
+                self.inner.put_user(&user)?;
+            }
+        }
+        self.permission_cache.map_write(|mut cache| {
+            cache.user_permissions.extend(new_perms.into_iter());
+            let _ignore = cache.role_to_users_map.remove(&req.role);
+        });
+        Ok(())
+    }
+
+    /// Sync `AuthRoleGrantPermissionRequest` and return whether authstore is changed.
+    fn sync_role_grant_permission_request(
+        &self,
+        req: &AuthRoleGrantPermissionRequest,
+    ) -> Result<(), ExecuteError> {
+        let mut role = self.inner.get_role(&req.name)?;
+        let permission = req
+            .perm
+            .clone()
+            .ok_or_else(ExecuteError::permission_not_given)?;
+
+        #[allow(clippy::indexing_slicing)] // this index is always valid
+        match role
+            .key_permission
+            .binary_search_by(|p| match p.key.cmp(&permission.key) {
+                Ordering::Equal => p.range_end.cmp(&permission.range_end),
+                Ordering::Less => Ordering::Less,
+                Ordering::Greater => Ordering::Greater,
+            }) {
+            Ok(idx) => {
+                role.key_permission[idx].perm_type = permission.perm_type;
+            }
+            Err(idx) => {
+                role.key_permission.insert(idx, permission.clone());
+            }
+        };
+        self.inner.put_role(&role)?;
+        self.permission_cache.map_write(move |mut cache| {
+            let users = cache
+                .role_to_users_map
+                .get(&req.name)
+                .cloned()
+                .unwrap_or_default();
+            let key_range = KeyRange::new(permission.key, permission.range_end);
+            for user in users {
+                let entry = cache
+                    .user_permissions
+                    .entry(user)
+                    .or_insert_with(UserPermissions::new);
+                #[allow(clippy::unwrap_used)] // safe unwrap
+                match Type::from_i32(permission.perm_type).unwrap() {
+                    Type::Readwrite => {
+                        entry.read.push(key_range.clone());
+                        entry.write.push(key_range.clone());
+                    }
+                    Type::Write => {
+                        entry.write.push(key_range.clone());
+                    }
+                    Type::Read => {
+                        entry.read.push(key_range.clone());
+                    }
+                }
+            }
+        });
+        Ok(())
+    }
+
+    /// Sync `AuthRoleRevokePermissionRequest` and return whether authstore is changed.
+    fn sync_role_revoke_permission_request(
+        &self,
+        req: &AuthRoleRevokePermissionRequest,
+    ) -> Result<(), ExecuteError> {
+        let mut role = self.inner.get_role(&req.role)?;
+        let idx = role
+            .key_permission
+            .binary_search_by(|p| match p.key.cmp(&req.key) {
+                Ordering::Equal => p.range_end.cmp(&req.range_end),
+                Ordering::Less => Ordering::Less,
+                Ordering::Greater => Ordering::Greater,
+            })
+            .map_err(|_ignore| ExecuteError::permission_not_granted())?;
+        let _ignore = role.key_permission.remove(idx);
+        self.inner.put_role(&role)?;
+        self.permission_cache.map_write(|mut cache| {
+            let users = cache
+                .role_to_users_map
+                .get(&req.role)
+                .map_or_else(Vec::new, |users| {
+                    users
+                        .iter()
+                        .filter_map(|user| self.inner.get_user(user).ok())
+                        .collect::<Vec<_>>()
+                });
+            for user in users {
+                let perms = self.get_user_permissions(&user);
+                let _old = cache
+                    .user_permissions
+                    .insert(String::from_utf8_lossy(&user.name).to_string(), perms);
+            }
+        });
+        Ok(())
     }
 
     /// Auth revision
@@ -81,7 +851,23 @@ where
         username: &str,
         password: &str,
     ) -> Result<i64, ExecuteError> {
-        self.inner.check_password(username, password)
+        if !self.is_enabled() {
+            return Err(ExecuteError::auth_not_enabled());
+        }
+        let user = self.inner.get_user(username)?;
+        let need_password = user.options.as_ref().map_or(true, |o| !o.no_password);
+        if !need_password {
+            return Err(ExecuteError::no_password_user());
+        }
+
+        let hash = String::from_utf8_lossy(&user.password);
+        let hash = PasswordHash::new(&hash)
+            .unwrap_or_else(|e| panic!("Failed to parse password hash, error: {e}"));
+        Pbkdf2
+            .verify_password(password.as_bytes(), &hash)
+            .map_err(|_ignore| ExecuteError::auth_failed())?;
+
+        Ok(self.revision())
     }
 
     /// Check if the request need admin permission
@@ -105,19 +891,24 @@ where
         )
     }
 
+    #[cfg(test)]
+    pub(super) fn permission_cache(&self) -> PermissionCache {
+        self.permission_cache.map_read(|cache| cache.clone())
+    }
+
     /// check if the request is permitted
     pub(crate) async fn check_permission(
         &self,
         wrapper: &RequestWithToken,
     ) -> Result<(), ExecuteError> {
-        if !self.inner.is_enabled() {
+        if !self.is_enabled() {
             return Ok(());
         }
         if let RequestWrapper::AuthenticateRequest(_) = wrapper.request {
             return Ok(());
         }
         let claims = match wrapper.token {
-            Some(ref token) => self.inner.verify_token(token)?,
+            Some(ref token) => self.verify_token(token)?,
             None => {
                 // TODO: some requests are allowed without token when auth is enabled
                 return Err(ExecuteError::token_not_provided());
@@ -265,7 +1056,7 @@ where
 
     /// check if user can revoke lease
     async fn check_lease(&self, username: &str, lease_id: i64) -> Result<(), ExecuteError> {
-        let lease = self.inner.get_lease(lease_id).await;
+        let lease = self.get_lease(lease_id).await;
         if let Some(lease) = lease {
             let keys = lease.keys();
             for key in keys {
@@ -277,7 +1068,7 @@ where
 
     /// Check if the user has admin permission
     fn check_admin_permission(&self, username: &str) -> Result<(), ExecuteError> {
-        if !self.inner.is_enabled() {
+        if !self.is_enabled() {
             return Ok(());
         }
         let user = self.inner.get_user(username)?;
@@ -300,12 +1091,47 @@ where
             return Ok(());
         }
         let key_range = KeyRange::new(key, range_end);
-        self.inner.check_permission(username, &key_range, perm_type)
+        self.check_permission_helper(username, &key_range, perm_type)
+    }
+
+    /// Check permission by username and key range
+    fn check_permission_helper(
+        &self,
+        username: &str,
+        key_range: &KeyRange,
+        permission_type: Type,
+    ) -> Result<(), ExecuteError> {
+        if let Some(permissions) = self.permission_cache.read().user_permissions.get(username) {
+            match permission_type {
+                Type::Read => {
+                    if permissions
+                        .read
+                        .iter()
+                        .any(|kr| kr.contains_range(key_range))
+                    {
+                        return Ok(());
+                    }
+                }
+                Type::Write => {
+                    if permissions
+                        .write
+                        .iter()
+                        .any(|kr| kr.contains_range(key_range))
+                    {
+                        return Ok(());
+                    }
+                }
+                Type::Readwrite => {
+                    unreachable!("Readwrite is unreachable");
+                }
+            }
+        }
+        Err(ExecuteError::PermissionDenied)
     }
 
     /// Assign root token
     pub(crate) fn root_token(&self) -> Result<String, ExecuteError> {
-        self.inner.assign(ROOT_USER)
+        self.assign(ROOT_USER)
     }
 }
 
@@ -345,7 +1171,7 @@ mod test {
         );
         assert!(exe_and_sync(&store, &req).is_ok());
         assert_eq!(
-            store.inner.permission_cache(),
+            store.permission_cache(),
             PermissionCache {
                 user_permissions: HashMap::from([(
                     "u".to_owned(),
@@ -372,7 +1198,7 @@ mod test {
         );
         assert!(exe_and_sync(&store, &req).is_ok());
         assert_eq!(
-            store.inner.permission_cache(),
+            store.permission_cache(),
             PermissionCache {
                 user_permissions: HashMap::from([("u".to_owned(), UserPermissions::new())]),
                 role_to_users_map: HashMap::from([("r".to_owned(), vec!["u".to_owned()])]),
@@ -391,7 +1217,7 @@ mod test {
         );
         assert!(exe_and_sync(&store, &req).is_ok());
         assert_eq!(
-            store.inner.permission_cache(),
+            store.permission_cache(),
             PermissionCache {
                 user_permissions: HashMap::from([("u".to_owned(), UserPermissions::new(),)]),
                 role_to_users_map: HashMap::new(),
@@ -410,7 +1236,7 @@ mod test {
         );
         assert!(exe_and_sync(&store, &req).is_ok());
         assert_eq!(
-            store.inner.permission_cache(),
+            store.permission_cache(),
             PermissionCache {
                 user_permissions: HashMap::new(),
                 role_to_users_map: HashMap::from([("r".to_owned(), vec![])]),
@@ -467,7 +1293,7 @@ mod test {
         );
         assert!(exe_and_sync(&store, &req4).is_ok());
         assert_eq!(
-            store.inner.permission_cache(),
+            store.permission_cache(),
             PermissionCache {
                 user_permissions: HashMap::from([(
                     "u".to_owned(),
